Add a simple callback-based INI file parser
This will be used for parsing the configuration file when that is added.

The supported file format included single- and double-quoted value
strings, comments beginning with ';' and continuing to the end of the
line and line continuations via backslash escapes.  It is an error for a
key-value line to occur before any section headings.