completion: Handle commands using a path
The completion script uses "compgen -c" with a modified PATH to find
commands inside the chroot.  This does not work when attempting to
execute a command containing "/".  Detect this case and use the
chpersroot compgen wrapper to perform filename completion inside the
chroot.

A new mode is added to the compgen wrapper to perform filename
completion and drop any files that are not executable.